+++ src/modules/guestbook/schema.ts
@@ -12,11 +12,12 @@ export type GuestMessageTable = {
   replyText?: string | null;
   createdAt: Generated<Date>;
   updatedAt: ColumnType<Date, string | undefined, string | undefined>;
+  reviewedAt: ColumnType<Date | null, string | undefined, string | undefined>;
 };
 
 export type GuestMessage = Selectable<GuestMessageTable>;
 export type NewGuestMessage = Omit<
   Insertable<GuestMessageTable>,
-  "id" | "status" | "replyText" | "updatedAt"
+  "id" | "status" | "replyText" | "reviewedAt" | "updatedAt"
 >;
 export type GuestMessageUpdate = Updateable<GuestMessageTable>;